--- a/benchmark_test.go
+++ b/benchmark_test.go
@@ -15,6 +15,7 @@ import (
 var Rb *Bitmap
 
 func BenchmarkNewBitmap(b *testing.B) {
+	b.ReportAllocs()
 	for i := 0; i < b.N; i++ {
 		Rb = New()
 	}