@lemire 2018-02-10
Thanks... it looks promising.

The changes look significant, so I think we want someone like  @maciej or @glycerine to review and recommend these changes.

Meanwhile, I have a few questions...

1. Doesn't this break our public API? If not, why not?

2. To me, at least, your changes look mysterious so it makes me nervous. If you could explain them briefly (as in, how does it work), that would help me.

3. What does this do...

```
var _ interface {
 	encoding.BinaryUnmarshaler
} = (*Bitmap)(nil)
``` 

It might be the first time I come across Go code that looks so mysterious.

4. You are changing the calling convention so that the roaring structs are passed by value rather than by pointer. Why?