@anacrolix 2018-02-10
1. Yes. However it's not too complicated as inferred types will continue to work, so mainly just `gofmt -r '*Bitmap->Bitmap'` with some luck. It's definitely worth checking against some of your bigger users, there are only a handful of [public ones](https://godoc.org/github.com/RoaringBitmap/roaring?importers).
2. The goal is that instantiating a Bitmap does not require a heap allocation.
3. It fails to compile if a *Bitmap does not implement the interface. It's a pattern that allows bundling a whole bunch of interfaces together for compile time checking.
4. To avoid forcing a heap allocation for ephemeral bitmaps, which [occurs a lot](https://github.com/anacrolix/torrent/blob/master/connection.go#L624) in my use case in anacrolix/torrent.

It's worth noting that there are other potential optimizations available now that the root type is not forced onto the heap. You can make highlowcontainer a pointer at a later date to regain reference behaviour, reduce memory footprint, and reduce overhead on a lot of the Bitmap methods. Also since the type is not a pointer, you can shift to an interface, or type alias back on to a pointer type in the future.