@lemire 2018-02-10
Could we avoid breaking the public API by creating a new struct type? It seems that it could be done with very little new code.