@anacrolix 2018-02-12
That's a great idea, give me a while to give it a try.