@lemire 2018-02-12
Thanks. We shall wait on your feedback.