@maciej 2018-03-04
I think that in light of Russ Cox's recent post on [Semantic Import Versioning](https://research.swtch.com/vgo-import) we should avoid breaking the public API. The post describes how in the Go ecosystem libraries should introduce backward-incompatible changes. I think Go roaring should be a good Go ecosystem citizen and abide by those rules. :-)