@lemire 2018-03-06
@maciej Speaking for myself, I take very seriously API breaking changes. The blog post you point to only reinforces my own belief. In this particular case, I think we can achieve the desired result without breaking the API.