@maciej 2019-08-13
@lemire perhaps it would make sense to close PRs due to inactivity so they don't linger around in the _open PR_ list. What do you think?