@lemire 2019-08-13
Transformed into an issue 

https://github.com/RoaringBitmap/roaring/issues/228